requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
passlib
jinja2
python-jose
SQLAlchemy
langchain
langchain-google-genai
pydantic
starlette
python-multipart
bcrypt


requests
langchain-core
PyAudio
soundfile
groq
elevenlabs
cloudinary